+++ crates/languages/src/json.rs
@@ -234,14 +234,8 @@ impl JsonLspAdapter {
         schemas
             .as_array_mut()
             .unwrap()
-            .extend(cx.all_action_names().iter().map(|name| {
-                let mut file_name = name.replace("::", "__");
-                let name_len = file_name.len();
-                file_name.push_str(".json");
-                serde_json::json!({
-                    "fileMatch": [file_name],
-                    "url": format!("zed://schemas/action/{}", &file_name[..name_len])
-                })
+            .extend(cx.all_action_names().into_iter().map(|&name| {
+                project::lsp_store::json_language_server_ext::url_schema_for_action(name)
             }));
 
         // This can be viewed via `dev: open language server logs` -> `json-language-server` ->
